Add option to manually pay bond principal
Added a new menu item for bond assets to allow user to enter details for
a manual payment of bond principal and optionally allow the transaction
to be scheduled in the future.
Also fixed a bug that updated the bond principal value even when a
transaction was not executed and was pending due to missing data.